% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{ondri_palette}
\alias{ondri_palette}
\title{An ONDRI palette generator}
\usage{
ondri_palette(name, start = 1, end, n, is_discrete = TRUE)
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{Cohort1}, \code{Cohort2}}

\item{start}{Starting index of colours desired.
If omitted, starts at 1.}

\item{end}{Ending index of colours desired.
If omitted, ends at last colour in palette list.}

\item{n}{Number of colours desired for type="continuous".
If omitted, uses amount of discrete colours in range [start:end].}

\item{is_discrete}{Either TRUE or FALSE. Set to FALSE if you want
to automatically interpolate between colours.}
}
\value{
A vector of colours.
}
\description{
These are a handful of colour palettes standard within ONDRI.
}
\details{
This function heavily borrows from Karthik Ram's wesanderson package, and specifically
\url{https://github.com/karthik/wesanderson/blob/master/R/colors.R}
}
\examples{
ondri_palette("Cohort1")
ondri_palette("Cohort1", start = 3, end = 7) # Returns only the VCI, FTD, ADMCI, PD, and ALS colours
ondri_palette("Cohort2", n = 20, is_discrete = FALSE)

# If you need more colours than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colours
pal <- ondri_palette(name = "Cohort1", start = 3, n = 20, is_discrete = FALSE)
image(volcano, col = pal)

}
\keyword{colors}
