% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{ondri_palettes}
\alias{ondri_palettes}
\title{Complete list of ONDRI palettes}
\format{
An object of class \code{list} of length 2.
}
\usage{
ondri_palettes
}
\description{
Complete list of ONDRI palettes
}
\keyword{datasets}
