% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONDRIpalettes.R
\docType{package}
\name{ondricolors}
\alias{ondricolors}
\title{ONDRI Palettes}
\description{
A lightweight package for the Ontario Neurodegenerative Disease Research Initiative (ONDRI) color palettes.
Returns a palette object (vector of colors) similar to Wes Anderson Palettes (https://github.com/karthik/wesanderson).
The palette attributes certain colors to cohort labels.
}
\details{
ONDRI's color palettes
}
\examples{
\dontrun{
## quick visualization of palettes
## continuous example
 pal1 <- ondri_palette(name = "Cohort1", start = 3, n = 20, is_discrete = FALSE)
image(volcano, col = pal1)

## discrete example
pal2 <- ondri_palette(name = "Cohort2")
image(volcano, col = pal2)

## three toy examples
library(ondricolors)
data("toy_ONDRI_data")
## first in base R
## for the filled pchs
col2use <- ondri_palette("Cohort2")[as.character(toy_ONDRI_data$COHORT)]
plot(toy_ONDRI_data$X, toy_ONDRI_data$Y, col = col2use, pch = 20)

plot(toy_ONDRI_data$X, toy_ONDRI_data$Y, bg = col2use, pch = 21)

## now with ggplot2
library(ggplot2)
ggp_color <- ggplot(toy_ONDRI_data, aes(x = X, y = Y, color = COHORT)) +
  geom_point() +
  scale_color_manual(values = ondri_palette("Cohort2"))
ggp_color
}
}
