% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{print.palette}
\alias{print.palette}
\title{print.palette}
\usage{
\method{print}{palette}(x, ...)
}
\description{
print.palette
}
\details{
This function heavily borrows from Karthik Ram's wesanderson package, and specifically
\url{https://github.com/karthik/wesanderson/blob/master/R/colors.R}
}
