% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONDRIpalettes.R
\docType{data}
\name{toy_ONDRI_data}
\alias{toy_ONDRI_data}
\title{Toy ONDRI data}
\format{
A data.frame with four columns:
\describe{
\item{SUBJECT}{A vector of SUBJECT IDs}
\item{COHORT}{A vector of COHORT values; one of five ("ADMCI","ALS","FTD","PD","VCI")}
\item{X}{Some arbitrary values for the x-axis in a plot}
\item{Y}{Some arbitrary values for the y-axis in a plot}
}
}
\usage{
toy_ONDRI_data
}
\description{
A small toy data set to help illustrate how to use the ondricolors package
}
\keyword{datasets}
